app/frontend/requirements.txt

fastapi
streamlit
requests
numpy
Pillow


